Moving the old pyglet plotting module and adding a warning
To make place for the new interface the old plotting module was
moved to another location. It was noted in examples, documentation
and tests.

The location has changed but the Plot function is still imported
from the same location. It's a proxy object giving the Plot from
the old pyglet module, but it adds a warning about the change.
Importing the pyglet Plot from the new location does not give a
warning.
