+++ sympy/plotting/pygletplot/color_scheme.py
