+++ sympy/plotting/pygletplot/managed_window.py
