+++ sympy/plotting/pygletplot/plot.py
