+++ sympy/plotting/pygletplot/plot_axes.py
