+++ sympy/plotting/pygletplot/plot_camera.py
