+++ sympy/plotting/pygletplot/plot_controller.py
