+++ sympy/plotting/pygletplot/plot_curve.py
