+++ sympy/plotting/pygletplot/plot_interval.py
