+++ sympy/plotting/pygletplot/plot_mode.py
