+++ sympy/plotting/pygletplot/plot_mode_base.py
