+++ sympy/plotting/pygletplot/plot_modes.py
