+++ sympy/plotting/pygletplot/plot_object.py
