+++ sympy/plotting/pygletplot/plot_rotation.py
