+++ sympy/plotting/pygletplot/plot_surface.py
