+++ sympy/plotting/pygletplot/plot_window.py
