+++ sympy/plotting/pygletplot/util.py
